+++ src/vs/workbench/contrib/scrollLocking/browser/scrollLocking.ts
