+++ src/vs/workbench/workbench.common.main.ts
@@ -393,5 +393,5 @@ import 'vs/workbench/contrib/share/browser/share.contribution';
 import 'vs/workbench/contrib/accountEntitlements/browser/accountsEntitlements.contribution';
 
 // Synchronized Scrolling
-import 'vs/workbench/contrib/scrollLocking/browser/syncScroll.contribution';
+import 'vs/workbench/contrib/scrollLocking/browser/scrollLocking.contribution';
 //#endregion
